requirements.txt
# OpenDraft - Python Dependencies
# Minimal dependencies for maximum simplicity

# === LLM APIs ===
anthropic>=0.20.0
openai>=1.0.0
google-generativeai>=0.8.0  # Updated for Google Search grounding support

# === Citation Management ===
pybtex>=0.24.0              # BibTeX parsing/generation
citeproc-py>=0.6.0          # Citation formatting (APA, MLA, Chicago, IEEE)
PyYAML>=6.0.0               # YAML frontmatter parsing

# === Document Export ===
markdown>=3.5.0             # Markdown parsing
weasyprint>=60.0            # PDF generation (no LaTeX needed!)
python-docx>=1.0.0          # Word document generation

# === HTTP Client & Web Scraping ===
requests>=2.31.0            # For CrossRef API, GPTZero API, web scraping
beautifulsoup4>=4.12.0      # HTML parsing for title/metadata scraping
lxml>=4.9.0                 # Fast XML/HTML parser for BeautifulSoup

# === Utilities ===
python-dotenv>=1.0.0        # Load .env files
rich>=13.0.0                # Pretty terminal output (for scripts)
